Consolidate auth things in the auth package
