+++ mediator-app/src/main/java/net/wasdev/gameon/mediator/auth/GameOnHeaderAuthFilter.java
@@ -1,4 +1,4 @@
-package net.wasdev.gameon.mediator;
+package net.wasdev.gameon.mediator.auth;
 
 import java.io.IOException;
 import java.time.Instant;
@@ -9,14 +9,20 @@ import javax.ws.rs.client.ClientRequestContext;
 import javax.ws.rs.client.ClientRequestFilter;
 import javax.ws.rs.core.MultivaluedMap;
 
+import net.wasdev.gameon.mediator.Log;
+import net.wasdev.gameon.mediator.MapClient;
+
+/**
+ * Outbound Client request filter. Used by {@link MapClient} to authenticate with the map service.
+ */
 public class GameOnHeaderAuthFilter extends GameOnHeaderAuth implements ClientRequestFilter {
-    
+
     public GameOnHeaderAuthFilter(String userId, String secret) {
         super(secret,userId);
-        if (secret == null)       
+        if (secret == null)
             throw new RuntimeException("NULL secret");
     }
-    
+
     @Override
     public void filter(ClientRequestContext context) throws IOException {
         try {
